Move helper methods to airflow.utils.operator_helpers
