+++ airflow/sensors/python.py
@@ -19,7 +19,7 @@ from typing import Callable, Dict, List, Optional
 
 from airflow.sensors.base_sensor_operator import BaseSensorOperator
 from airflow.utils.decorators import apply_defaults
-from airflow.utils.helpers import determine_kwargs
+from airflow.utils.operator_helpers import determine_kwargs
 
 
 class PythonSensor(BaseSensorOperator):
